+++ packages/dataviews/src/index.js
@@ -1,3 +1,3 @@
 export { default as DataViews } from './dataviews';
-export { VIEW_LAYOUTS } from './constants';
+export { VIEW_LAYOUTS } from './layouts';
 export { filterSortAndPaginate } from './filter-and-sort-data-view';
