+++ packages/dataviews/src/normalize-fields.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import type { Field } from './types';
+import type { Field, NormalizedField } from './types';
 
 /**
  * Apply default values and normalize the fields config.
@@ -9,12 +9,13 @@ import type { Field } from './types';
  * @param fields Fields config.
  * @return Normalized fields config.
  */
-export function normalizeFields( fields: Field[] ): Field[] {
+export function normalizeFields( fields: Field[] ): NormalizedField[] {
 	return fields.map( ( field ) => {
 		const getValue = field.getValue || ( ( { item } ) => item[ field.id ] );
 
 		return {
 			...field,
+			header: field.header || field.id,
 			getValue,
 			render: field.render || getValue,
 		};
